refactor(switch): consolidate handle_switch into worktree/switch
`handle_switch.rs` was a thin coordinator over `worktree/switch.rs` —
loading config, planning, approving hooks, executing, and handling
output/shell-integration prompts. The split forced cross-file imports of
`SwitchOptions`, `SwitchPlan`, `SwitchBranchInfo`, `SwitchResult`,
`plan_switch`, and `execute_switch` despite all the real work living in
`worktree/switch.rs`.

Move the body into `worktree/switch.rs` as `run_switch` (renamed from
`handle_switch`) and delete `src/commands/handle_switch.rs`. Update the
two call sites (`main.rs`, `commands/picker/mod.rs`) and the
`STDOUT_ALLOWED_PATHS` allowlist. Drop the now-unused `SwitchPlan`
re-export from `commands/worktree/mod.rs`. No behavior change.
