+++ src/git/error.rs
@@ -236,7 +236,7 @@ impl SwitchSuggestionCtx {
     /// separator when this context is applied.
     fn apply(&self, cmd: String) -> String {
         let mut result = cmd;
-        // Flags are pre-escaped at construction (handle_switch.rs uses shell_escape)
+        // Flags are pre-escaped at construction (worktree/switch.rs uses shell_escape)
         for flag in &self.extra_flags {
             result.push(' ');
             result.push_str(flag);
